test(e2e): harden settings CRUD fixtures and Playwright stability
